fix: Update frontend Dockerfile to use Node 20 for Vite 7 compatibility
- Changed from Node 18 to Node 20 in all stages of Dockerfile.frontend
- Fixed crypto.hash error in Vite development server
- Removed user specification from database service in docker-compose.yml
- Updated Dockerfile.db to not copy SQL scripts (using Alembic instead)
