--- alembic/versions/e1ec726d02c6_add_last_login_to_users.py
@@ -1,103 +1,0 @@
-"""add_last_login_to_users
-
-Revision ID: e1ec726d02c6
-Revises: 
-Create Date: 2025-01-16 08:13:06.574447
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'e1ec726d02c6'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('user_roles')
-    op.drop_table('roles')
-    op.add_column('organizations', sa.Column('description', sa.String(), nullable=True))
-    op.drop_constraint('organizations_name_key', 'organizations', type_='unique')
-    op.create_index(op.f('ix_organizations_id'), 'organizations', ['id'], unique=False)
-    op.create_index(op.f('ix_organizations_name'), 'organizations', ['name'], unique=False)
-    op.add_column('sites', sa.Column('access_token', sa.String(), nullable=True))
-    op.add_column('sites', sa.Column('access_secret', sa.String(), nullable=True))
-    op.alter_column('sites', 'url',
-               existing_type=sa.TEXT(),
-               type_=sa.String(),
-               existing_nullable=False)
-    op.alter_column('sites', 'description',
-               existing_type=sa.TEXT(),
-               type_=sa.String(),
-               existing_nullable=True)
-    op.alter_column('sites', 'organization_id',
-               existing_type=sa.INTEGER(),
-               nullable=False)
-    op.drop_constraint('sites_name_key', 'sites', type_='unique')
-    op.create_index(op.f('ix_sites_id'), 'sites', ['id'], unique=False)
-    op.create_index(op.f('ix_sites_name'), 'sites', ['name'], unique=False)
-    op.create_unique_constraint(None, 'sites', ['url'])
-    op.add_column('users', sa.Column('last_login', sa.DateTime(), nullable=True))
-    op.alter_column('users', 'hashed_password',
-               existing_type=sa.TEXT(),
-               type_=sa.String(),
-               existing_nullable=False)
-    op.drop_constraint('users_email_key', 'users', type_='unique')
-    op.create_index(op.f('ix_users_email'), 'users', ['email'], unique=True)
-    op.create_index(op.f('ix_users_id'), 'users', ['id'], unique=False)
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_index(op.f('ix_users_id'), table_name='users')
-    op.drop_index(op.f('ix_users_email'), table_name='users')
-    op.create_unique_constraint('users_email_key', 'users', ['email'])
-    op.alter_column('users', 'hashed_password',
-               existing_type=sa.String(),
-               type_=sa.TEXT(),
-               existing_nullable=False)
-    op.drop_column('users', 'last_login')
-    op.drop_constraint(None, 'sites', type_='unique')
-    op.drop_index(op.f('ix_sites_name'), table_name='sites')
-    op.drop_index(op.f('ix_sites_id'), table_name='sites')
-    op.create_unique_constraint('sites_name_key', 'sites', ['name'])
-    op.alter_column('sites', 'organization_id',
-               existing_type=sa.INTEGER(),
-               nullable=True)
-    op.alter_column('sites', 'description',
-               existing_type=sa.String(),
-               type_=sa.TEXT(),
-               existing_nullable=True)
-    op.alter_column('sites', 'url',
-               existing_type=sa.String(),
-               type_=sa.TEXT(),
-               existing_nullable=False)
-    op.drop_column('sites', 'access_secret')
-    op.drop_column('sites', 'access_token')
-    op.drop_index(op.f('ix_organizations_name'), table_name='organizations')
-    op.drop_index(op.f('ix_organizations_id'), table_name='organizations')
-    op.create_unique_constraint('organizations_name_key', 'organizations', ['name'])
-    op.drop_column('organizations', 'description')
-    op.create_table('roles',
-    sa.Column('id', sa.INTEGER(), server_default=sa.text("nextval('roles_id_seq'::regclass)"), autoincrement=True, nullable=False),
-    sa.Column('name', sa.VARCHAR(length=255), autoincrement=False, nullable=False),
-    sa.Column('description', sa.TEXT(), autoincrement=False, nullable=True),
-    sa.PrimaryKeyConstraint('id', name='roles_pkey'),
-    sa.UniqueConstraint('name', name='roles_name_key'),
-    postgresql_ignore_search_path=False
-    )
-    op.create_table('user_roles',
-    sa.Column('user_id', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.Column('role_id', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.ForeignKeyConstraint(['role_id'], ['roles.id'], name='user_roles_role_id_fkey'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], name='user_roles_user_id_fkey'),
-    sa.PrimaryKeyConstraint('user_id', 'role_id', name='user_roles_pkey')
-    )
-    # ### end Alembic commands ###
