--- alembic/versions/f2f02d637f4c_add_organization_id_and_role_to_users.py
@@ -1,50 +1,0 @@
-"""add_organization_id_and_role_to_users
-
-Revision ID: f2f02d637f4c
-Revises: e1ec726d02c6
-Create Date: 2025-02-22 12:12:42.880001
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import postgresql
-
-# revision identifiers, used by Alembic.
-revision: str = 'f2f02d637f4c'
-down_revision: Union[str, None] = 'e1ec726d02c6'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('users', sa.Column('organization_id', sa.Integer(), nullable=True))
-    op.add_column('users', sa.Column('role', sa.String(), nullable=True))
-    op.alter_column('users', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               nullable=False,
-               existing_server_default=sa.text('CURRENT_TIMESTAMP'))
-    op.alter_column('users', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               nullable=False,
-               existing_server_default=sa.text('CURRENT_TIMESTAMP'))
-    op.create_foreign_key(None, 'users', 'organizations', ['organization_id'], ['id'])
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'users', type_='foreignkey')
-    op.alter_column('users', 'updated_at',
-               existing_type=postgresql.TIMESTAMP(),
-               nullable=True,
-               existing_server_default=sa.text('CURRENT_TIMESTAMP'))
-    op.alter_column('users', 'created_at',
-               existing_type=postgresql.TIMESTAMP(),
-               nullable=True,
-               existing_server_default=sa.text('CURRENT_TIMESTAMP'))
-    op.drop_column('users', 'role')
-    op.drop_column('users', 'organization_id')
-    # ### end Alembic commands ###
